Changed signature of read() in OwDevice interface, to generalize read() to all sensor outputs.
Now read(Readings r) takes an enum value (defined in OwDevice) representing the readings type the OwDevice impl should return.

Single value sensors can ignore this (or throw an appropriate error if passed the wrong argument), and multiple value sensors can use this to determine the reading to return.

Consurmers of OwDevice implementations are now closer to only having to know the interface. Before, they needed to know the api of the actual sensor type implemenation.

This is first step in a rewrite of the OwDevice hierarchy to better decouple consumers' awareness of the actual implementations.

Relevant changes also to unit tests, and getUnit() that now works similar to read().